Make Rule.monitor 1-based instead of 0-based
In the other fields of the Rule struct, zero or NULL values signify
"match any" or "use default."  For the monitor field, this was -1.  If
we number monitors starting with 1, then 0 is available for this
purpose.

This is extra convenient because 0/NULL is the default value when using
static allocation or struct initializers, so now "field not present"
gives the default behavior for all fields.